requirements.txt
pymilvus>=2.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
pandas>=2.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
